Fix copying records between codeplugs
This used to fail.  For example, if copying a zone, and the channels
did not exist, the copy would fail.  We now also copy all of the
records the requested records depend on.

Undo/redo of this operation does not preserve the order of the records.
I'm still looking into this.  I'm tempted to remove undo/redo entirely.